ws_client/requirements.txt
websocket-client==1.6.2
mariadb==1.1.14